third_party/blink/web_tests/platform/mac-mac10.14/external/wpt/WebIDL/ecmascript-binding/sequence-conversion-expected.txt

This is a testharness.js-based test.
PASS An array
PASS A generator
FAIL An array with an overridden Symbol.iterator assert_array_equals: expected property 0 to be 6 but got 1 (expected array [6, 7] got [1, 2])
PASS An object with an overriden Symbol.iterator
FAIL An array with an overridden Symbol.iterator on the prototype assert_array_equals: value is 11, expected array
FAIL An array with an overridden %ArrayIterator%.prototype.next Maximum call stack size exceeded
PASS A holey array with fallback to an accessor on the prototype
PASS A string array in sequence<sequence> or record
Harness: the test ran to completion.
